Telus_flourish/requirements.txt
streamlit
numpy
opencv-python
Pillow
torch
torchvision
requests
